add yaml support, requests init options, adhoc redcap options
This commit adds a lot of quaility of life features. Yaml versions are
easier to handle, but we don't break the existing json versions. Also
requests sometimes needs particular options to make it work a certain
way dependent on the network. Adhoc redcap and master.yaml version
added to facilitate most use cases